requirements.txt
Django==5.2.11
django-admin-sortable2==2.3.1
django-appconf==1.2.0
django-environ==0.12.0
django-imagekit==6.0.0
django-jazzmin==3.0.2
django-modeltranslation==0.19.19
gunicorn==25.1.0
pilkit==3.0
pillow==12.1.1
psycopg[binary]==3.2.3
python-slugify==8.0.4
sqlparse==0.5.5